ticket-service/requirements.txt
fastapi
uvicorn
redis
pydantic
psycopg2-binary
pytest
httpx